Update Factorial
In this version, I used the memo dictionary to store the results of previously computed factorials. If a value has already been computed, we simply look it up in the memo dictionary instead of recalculating it, which can significantly improve the performance for larger values of n.

With this optimization, the program will be more efficient when calculating factorials using recursion.